% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPR_functions.R
\name{bioProp_EWR_partialCor_adjustment}
\alias{bioProp_EWR_partialCor_adjustment}
\title{Title: EWR using partial covariance matrix for stage 2, faster option}
\usage{
bioProp_EWR_partialCor_adjustment(
  X,
  Y,
  Z_EWR,
  Z_proportion_ratio = NULL,
  bootstrapping_number = 100,
  covClumpThreshold = 0.5
)
}
\arguments{
\item{X}{a vector which denote the levels of the target protein.}

\item{Y}{a vector (same length of X) which denote the levels of the helper protein, should be significantly correlated with X.}

\item{Z_EWR}{for performing EWR regression: a matrix (each row correspond to one element in X and Y) of auxiliary proteins that tags biological variance in X. We recommend at least 100 columns.}

\item{Z_proportion_ratio}{for estimate ratio of variance: a matrix (each row correspond to one element in X and Y) of auxiliary proteins that tags biological variance in X. We recommend at least 100 columns.}

\item{bootstrapping_number}{number of bootstraps for computing confidence interval.}

\item{covClumpThreshold}{when computing the partial correlation matrix, clumping highly correlated pairs of Zs; default threshold cor>0.5.}
}
\value{
a list containing objects:
\itemize{
\item Bioprop_xi1_xi2: estimate of biological proportion (i.e. 1-noise) variance in protein level X.
\item sampling_se: standard error of the estimate Bioprop_xi1_xi2; estimated through sampling procedure.
\item sampling_mean: mean across the sample estimate (could be used as a bias detection tool).
\item jk_Zlabel_se_bioprop: se estimated by jacknifing Zj j index (as opposed to jackknifing individuals).
\item PxPy
\item PxPy_se: standard error of PxPy estimate.
\item beta_EWR: Slope of covariance y2z on x2z.
\item beta_EWR_se: standard error of beta_EWR estimate.
\item Bioprop_xi1: xi_1 variance
\item Partial_beta: partial causal effect
\item Bioprop_Y: beta^2 * E_xi12
\item sample_Bioprop_xi1: sample mean of Bioprop_xi1
\item sample_Partial_beta: sample mean of sample_Partial_beta
\item sample_Bioprop_Y: sample mean of Bioprop_Y
\item adjusting_factor: adjusting factor computed from the correlation matrix.
\item sample_adjusting_factor: adjusting factor computed from the correlation matrix, mean across bootstrap samples.
\item E_xi_XY: estimated variance in X that is shared with Y from the correlation matrix.
\item E_xi_X: estimated variance in X that is not shared with Y from the correlation matrix.
\item sample_E_xi_XY: bootstrap mean of E_xi_XY.
\item sample_E_xi_X: bootstrap mean of E_xi_X.
\item X_variance_corMatrix: proportion of variance estimated directly from the Z.
\item se_X_variance_corMatrix: bootstrap se of X_variance_corMatrix.
\item sigma_X: directly estimated residual variance in X that is not captured by Z.
\item sigma_Y: directly estimated residual variance in Y that is not captured by Z.
}
}
\description{
Title: EWR using partial covariance matrix for stage 2, faster option
}
\examples{
quick_results <- bioProp_EWR_partialCor_adjustment(X_EXAMPLE, Y_EXAMPLE, Z_EWR = Z_EXAMPLE[,1:200], bootstrapping_number=20)

}
