% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPR_functions.R
\name{BPR_fast_wrapper}
\alias{BPR_fast_wrapper}
\title{Title: BPR wrapper function. Input is many pairs of X and Y, and all based on the same set of Z.}
\usage{
BPR_fast_wrapper(
  target_protein,
  auxiliary_proteins,
  top_PCs = 50,
  number_of_Y_chosen = 5,
  bootstrapping_number = 100,
  threshold_PxPy = 0,
  Z_pca_number = 500
)
}
\arguments{
\item{target_protein}{a vector which denote the levels of the target protein.}

\item{auxiliary_proteins}{matrix (each row correspond to one element in target proteins) of auxiliary proteins that tags biological variance in target proteins. We recommend at least 100 columns.}

\item{top_PCs}{number of PCs estimated from auxiliary proteins; default 20.}

\item{number_of_Y_chosen}{number of PCs used as Y for the EWR estimation; deafalt 5.}

\item{bootstrapping_number}{number of boostrap samples for estimate uncertainty.}

\item{threshold_PxPy}{a threshold for cor(PC, X)^2, default 0.}

\item{Z_pca_number}{number of Ys chosen to compute Y.}
}
\value{
A dataframe contains same output as the \code{bioProp_EWR_partialCor_adjustment} function.
}
\description{
Title: BPR wrapper function. Input is many pairs of X and Y, and all based on the same set of Z.
}
\examples{
full_run_results_big_wrapper <- BPR_fast_wrapper(X_EXAMPLE,Z_EXAMPLE,number_of_Y_chosen = 2, bootstrapping_number = 5)
}
