% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPR_functions.R
\name{IVW_summary}
\alias{IVW_summary}
\title{Title: Inverse variance weighting combines different estimates}
\usage{
IVW_summary(bioProp_results)
}
\arguments{
\item{bioProp_results}{output of  \code{EWR} function.}
}
\value{
Summarised dataframe with IVW estimate.
}
\description{
Title: Inverse variance weighting combines different estimates
}
\examples{
  full_run_results_big_wrapper <- EWR(X_EXAMPLE,Z_EXAMPLE, use_as_helper = 2)
  IVW_summary(full_run_results_big_wrapper)
}
