% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPR_functions.R
\name{BPR_fast_wrapper}
\alias{BPR_fast_wrapper}
\title{Title: BPR wrapper function. Input is many pairs of X and Y, and all based on the same set of Z.
Paralle computation of all bootstrap samples to speed up the computation; this function is much faster for many pairs of X and Y.}
\usage{
BPR_fast_wrapper(
  target_traits,
  helper_traits,
  auxiliary_traits,
  bootstrapping_number = 20,
  covClumpThreshold = 0.5
)
}
\arguments{
\item{target_traits}{a matrix (nrow= number of samples; ncolumns=number of target traits), each column is one target trait whose biological proportion is to be estimated.}

\item{helper_traits}{a matrix of the same size as target_traits; each column is the helper trait for the corresponding target trait; column index should match.}

\item{auxiliary_traits}{matrix (each row correspond to one element in target proteins) of auxiliary proteins that tags biological variance in target proteins. We recommend at least 100 columns.}

\item{bootstrapping_number}{number of boostrap samples for estimate uncertainty.}

\item{covClumpThreshold}{when computing the partial correlation matrix, clumping highly correlated pairs of Zs; default threshold cor>0.5.}
}
\value{
A dataframe contains same output as the \code{bioProp_BPR_partialCor_adjustment} function.
}
\description{
Title: BPR wrapper function. Input is many pairs of X and Y, and all based on the same set of Z.
Paralle computation of all bootstrap samples to speed up the computation; this function is much faster for many pairs of X and Y.
}
\examples{
full_run_results_big_wrapper <- BPR_fast_wrapper(cbind(X_EXAMPLE,X_EXAMPLE, X_EXAMPLE), cbind(Y_EXAMPLE,Y_EXAMPLE, Z_EXAMPLE[,1]), Z_EXAMPLE, bootstrapping_number = 5)
}
